pyproject.toml
[project]
name = "lovdtools"
description = "A fluent API client for the global shared LOVD instance."
dynamic = [
  "authors",
  "version"
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "platformdirs>=4.3.8",
    "polars>=1.32.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "returns>=0.26.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/lovd/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/lovd"]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
]
